% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{engine_version}
\alias{engine_version}
\title{Bundled SparseDiffEngine version}
\usage{
engine_version()
}
\value{
A character scalar, e.g. \code{"0.6.1"}.
}
\description{
Returns the version string of the SparseDiffEngine C library bundled with
this package.
}
\examples{
engine_version()
}
