% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsediff-package.R
\docType{package}
\name{sparsediff-package}
\alias{sparsediff}
\alias{sparsediff-package}
\title{sparsediff: R interface to the SparseDiffEngine differentiation backend}
\description{
R bindings to the 'SparseDiffEngine' C library --- the sparse Jacobian and
Hessian differentiation backend used by 'CVXPY' for its Disciplined
Nonlinear Programming (DNLP) extension. This package is the R analog of the
'sparsediffpy' Python package and wraps the same C library (pinned at the
upstream v0.6.1 release).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bnaras.github.io/sparsediff/}
  \item \url{https://github.com/bnaras/sparsediff}
  \item Report bugs at \url{https://github.com/bnaras/sparsediff/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu}

Authors:
\itemize{
  \item Balasubramanian Narasimhan \email{naras@stanford.edu}
  \item Daniel Cederberg (Author of the bundled SparseDiffEngine C library) [copyright holder]
  \item William Zijie Zhang (Author of the bundled SparseDiffEngine C library) [copyright holder]
}

}
\keyword{internal}
