% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsediff-atoms.R
\name{sparsediff-bivariate}
\alias{sparsediff-bivariate}
\alias{sd_elementwise_mult}
\alias{sd_matmul}
\alias{sd_quad_over_lin}
\alias{sd_rel_entr}
\alias{sd_rel_entr_first_scalar}
\alias{sd_rel_entr_second_scalar}
\title{Bivariate atoms}
\usage{
sd_elementwise_mult(l, r)
sd_matmul(x, y)
sd_quad_over_lin(l, r)
sd_rel_entr(l, r)
sd_rel_entr_first_scalar(l, r)
sd_rel_entr_second_scalar(l, r)
}
\arguments{
\item{l, r, x, y}{expression handles.}
}
\value{
An expression handle.
}
\description{
Functions of two expression arguments.
}
\details{
\describe{
  \item{\code{sd_elementwise_mult}}{elementwise (Hadamard) product.}
  \item{\code{sd_matmul}}{matrix product \eqn{x y}.}
  \item{\code{sd_quad_over_lin}}{the quadratic-over-linear \eqn{\lVert x \rVert^2 / y}.}
  \item{\code{sd_rel_entr}}{elementwise relative entropy \eqn{x \log(x / y)}.}
  \item{\code{sd_rel_entr_first_scalar}, \code{sd_rel_entr_second_scalar}}{relative
    entropy with a scalar first or second argument broadcast against the other.}
}
}
\seealso{
\code{\link{sparsediff-elementwise}}, \code{\link{sparsediff-reduction}}
}
