% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsediff-atoms.R
\name{sparsediff-elementwise}
\alias{sparsediff-elementwise}
\alias{sd_exp}
\alias{sd_log}
\alias{sd_sin}
\alias{sd_cos}
\alias{sd_tan}
\alias{sd_sinh}
\alias{sd_tanh}
\alias{sd_asinh}
\alias{sd_atanh}
\alias{sd_logistic}
\alias{sd_xexp}
\alias{sd_normal_cdf}
\alias{sd_entr}
\alias{sd_power}
\alias{sd_neg}
\title{Elementwise atoms}
\arguments{
\item{child, c}{an expression handle (the argument).}

\item{p}{exponent for \code{sd_power}.}
}
\value{
An expression handle.
}
\description{
Smooth elementwise functions of a single expression. Each returns an
expression of the same shape as its argument.
}
\details{
\describe{
  \item{\code{sd_exp}, \code{sd_log}}{exponential and natural logarithm.}
  \item{\code{sd_sin}, \code{sd_cos}, \code{sd_tan}}{trigonometric functions.}
  \item{\code{sd_sinh}, \code{sd_tanh}, \code{sd_asinh}, \code{sd_atanh}}{hyperbolic and inverse-hyperbolic functions.}
  \item{\code{sd_logistic}}{the logistic \eqn{\log(1 + e^x)}.}
  \item{\code{sd_xexp}}{\eqn{x e^x}.}
  \item{\code{sd_normal_cdf}}{the standard normal CDF.}
  \item{\code{sd_entr}}{the elementwise entropy \eqn{-x \log x}.}
  \item{\code{sd_power}}{the power \eqn{x^p}.}
  \item{\code{sd_neg}}{negation \eqn{-x}.}
}
}
\seealso{
\code{\link{sparsediff-affine}}, \code{\link{sparsediff-bivariate}}
}
