% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsediff-atoms.R
\name{sparsediff-leaves}
\alias{sparsediff-leaves}
\alias{sd_variable}
\alias{sd_parameter}
\title{Leaf expressions: variables and parameters}
\arguments{
\item{d1, d2}{row and column dimensions of the leaf.}

\item{var_id}{0-based flat (column-major) offset of this variable's first
entry within the primal vector \code{u}.}

\item{param_id}{0-based parameter offset, analogous to \code{var_id}.}

\item{n_vars}{total number of variables in the problem.}

\item{values}{numeric data for the parameter (length \code{d1 * d2},
column-major).}
}
\value{
An expression handle.
}
\description{
Create the leaves of an expression graph. A variable is a slice of the
differentiation vector; a parameter is fixed data that can be updated between
evaluations (see \code{\link{sd_register_params}}).
}
\seealso{
\code{\link{sparsediff-elementwise}}, \code{\link{sd_problem}}
}
