% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsediff-atoms.R
\name{sparsediff-matrix}
\alias{sparsediff-matrix}
\alias{sd_scalar_mult}
\alias{sd_vector_mult}
\alias{sd_convolve}
\alias{sd_quad_form}
\alias{sd_left_matmul}
\alias{sd_right_matmul}
\alias{sd_left_matmul_dense}
\alias{sd_right_matmul_dense}
\title{Parameter- and constant-matrix atoms}
\arguments{
\item{param}{a parameter expression handle (see \code{\link{sd_parameter}}).}

\item{child}{an expression handle (the variable argument).}

\item{Qp, Qi, Qx}{the column-pointer, row-index and value arrays of a
compressed-sparse-column matrix \eqn{Q} (as in a \code{Matrix::dgCMatrix}:
\code{@p}, \code{@i}, \code{@x}) for \code{sd_quad_form}'s \eqn{x^\top Q x}.}

\item{Ap, Ai, Ax}{the compressed-sparse-column arrays of a constant matrix
\eqn{A} for the sparse matrix products.}

\item{ncol}{number of columns of the sparse constant matrix \eqn{A}.}

\item{m, n}{row and column dimensions of the dense constant matrix.}

\item{data}{the dense constant-matrix entries (length \code{m * n},
column-major).}
}
\value{
An expression handle.
}
\description{
Operations that combine an expression with fixed data --- a registered
parameter node or a constant matrix --- so the data can flow through the
differentiated graph (and, for parameters, be updated between evaluations).
}
\details{
\describe{
  \item{\code{sd_scalar_mult}, \code{sd_vector_mult}}{multiply a child by a
    scalar / vector parameter.}
  \item{\code{sd_convolve}}{convolution of a parameter kernel with a child.}
  \item{\code{sd_quad_form}}{the quadratic form \eqn{x^\top Q x} with sparse
    constant \eqn{Q}.}
  \item{\code{sd_left_matmul}, \code{sd_right_matmul}}{left / right product
    with a sparse constant matrix \eqn{A}.}
  \item{\code{sd_left_matmul_dense}, \code{sd_right_matmul_dense}}{left / right
    product with a dense constant matrix.}
}
}
\seealso{
\code{\link{sd_parameter}}, \code{\link{sd_register_params}}
}
